+++ pytest.ini
@@ -1,5 +1,8 @@
 [pytest]
+python_files = test_*.py
 pythonpath = .
+addopts = --tb=short
+
 markers =
     positive: Positive test case
     negative: Negative test case
@@ -9,3 +12,5 @@ markers =
     optional_fields: Tests optional schema fields
     behavior: Behavior-based validation
 
+plugins =
+    infra.logging.pytest_hooks
